// src/residual-config/enums/device-type.ts
export enum DeviceType {
    WORKSTATION = 'Workstation',
    LAPTOP = 'Ordinateur portable',
    MONITOR = 'Ecran',
    SMARTPHONE = 'Smartphone',
    SERVER = 'Serveur',
    STORAGE_SYSTEM = 'Système de stockage',
    PRINTER = 'Imprimante',
    SCANNER = 'Scanner',
    NETWORK_EQUIPMENT = 'Équipement réseau',
    OTHER = 'Autre'
}